+++ README.md
@@ -22,16 +22,21 @@
 Редъюсеры разделены и собраны в src/components/srvices/reducers/index.js
 
 import { combineReducers } from 'redux';
-import { getDataReducer,  getOrderReducer} from './api';
-import { itemReducer, madalReducer} from './burger-ingredients';
+import { getDataReducer} from './data-reducer';
+import { getOrderReducer} from './order-reducer';
+import { itemReducer} from './item-reducer';
 import { ingredientMoveReducer} from './burger-constructor';
+import { madalReducer} from './modal-reducer';
 
 export const rootReducer = combineReducers({
   data: getDataReducer,  // запрос данных ингредиентов с сервера
   order: getOrderReducer, // запоос заказа с сервера
   item: itemReducer, // открытие модального окна с ингредиентом
-  open: madalReducer, // открытие модального окана с заказщь
+  open: madalReducer, // открытие модального окана с заказом
   ingredients: ingredientMoveReducer, // работа конструктора: перемещение инградиента в конструктор, сортировка инградиентов в конструкторе
 
 });
+
+});
 (commit 84)
+* Созданы раздельные директории под разные редьюсеры (commit 85)
