+++ src/components/services/reducers/order-reducer.js
@@ -1,44 +1,11 @@
 
 import {
-  GET_DATA_REQUEST,
-  GET_DATA_SUCCESS,
-  GET_DATA_FAILED,
   GET_ORDER_REQUEST,
   GET_ORDER_SUCCESS,
   GET_ORDER_FAILED,
 } from '../actions/api';
 
-const initialStateData = {
-  data: [],
-  dataRequest: false,
-  dataFailed: false,
-  
-};
 
-export const getDataReducer = (state = initialStateData, action) => {
-  switch (action.type) {
-    case GET_DATA_REQUEST: {
-      return {...state, 
-            dataRequest: true }
-    }
-    case GET_DATA_SUCCESS: {
-      return {...state, 
-            dataRequest: false,
-            dataFailed: false,
-            data: action.data
-            }
-    }
-     case GET_DATA_FAILED: {
-      return {...state, 
-            dataRequest: false,
-            dataFailed: true,
-            }
-    }
-    default: {
-      return state;
-    }
-  }
-};
 
 const initialStateOrder = {
   orderNumber: null,
